requirements.txt
fastapi
uvicorn
google-genai
pydantic
sqlalchemy
python-multipart
python-dotenv
